+++ client/src/adt/operations/AdtObjectActivator.ts
@@ -1,16 +1,29 @@
 import { ADTClient, isAdtError, inactiveObjectsInResults } from "abap-adt-api"
 import { AbapObject } from "../abap/AbapObject"
 import { IncludeLensP } from "./IncludeLens"
-import { Uri } from "vscode"
+import { Uri, EventEmitter } from "vscode"
+
+export interface ActivationEvent {
+  object: AbapObject
+  uri: Uri
+  activated: AbapObject
+  mainProg?: string
+}
 
 export class AdtObjectActivator {
   constructor(private client: ADTClient) {}
+  private emitter = new EventEmitter<ActivationEvent>()
+
+  public get onActivate() {
+    return this.emitter.event
+  }
 
   public async activate(object: AbapObject, uri: Uri) {
     // TODO: handle multiple inactive components
     const inactive = object.getActivationSubject()
     let result
     let message
+    let mainProg: string | undefined
     try {
       result = await this.client.activate(inactive.name, inactive.path)
       if (result.inactive.length > 0) {
@@ -20,7 +33,7 @@ export class AdtObjectActivator {
     } catch (e) {
       if (isAdtError(e) && e.type === "invalidMainProgram") {
         const provider = IncludeLensP.get()
-        const mainProg = await provider.selectIncludeIfNeeded(uri)
+        mainProg = await provider.selectIncludeIfNeeded(uri)
         if (mainProg)
           result = await this.client.activate(
             inactive.name,
@@ -30,6 +43,7 @@ export class AdtObjectActivator {
       } else message = e.toString()
     }
     if (result && result.success) {
+      this.emitter.fire({ object, uri, activated: inactive, mainProg })
       await inactive.loadMetadata(this.client)
     } else {
       message =