+++ client/src/lib/logger.ts
@@ -1,6 +1,8 @@
 import { window } from "vscode"
 
-export const channel = window.createOutputChannel("ABAP FS")
+export const CHANNELNAME = "ABAP FS"
+
+export const channel = window.createOutputChannel(CHANNELNAME)
 export function log(...messages: string[]) {
   channel.appendLine(messages.join(""))
 }