+++ client/src/views/abapunit/index.ts
@@ -1,0 +1,2 @@
+export { alertManagers } from "./alerts"
+export * from "./register"