+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Added
+
+- test explorer integraton
+
 ## 0.7.23 2020-04-06
 
 ### Added